[MOV] product,web_sale -> sale: move code to its rightful place
Also remove groups usage from portal templates, if there is
a discount, it should be shown to the customer (and there probably
won't be any if the group is not enabled anyway).
