+++ addons/product/security/product_security.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-<data noupdate="0">
 
     <record id="group_product_pricelist" model="res.groups">
         <field name="name">Basic Pricelists</field>
@@ -23,12 +22,6 @@
         <field name="category_id" ref="base.module_category_hidden"/>
     </record>
 
-    <record id="group_discount_per_so_line" model="res.groups">
-        <field name="name">Discount on lines</field>
-        <field name="category_id" ref="base.module_category_hidden"/>
-    </record>
-
-</data>
 <data noupdate="1">
 
     <record id="product_comp_rule" model="ir.rule">
