+++ addons/sale/models/sale_order.py
@@ -1,4 +1,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from collections import defaultdict
 from datetime import timedelta
 from itertools import groupby
