+++ addons/sale_management/report/sale_report_templates.xml
@@ -12,7 +12,7 @@
                 <thead class="fw-bold">
                     <tr>
                         <td name="th_option_name" class="text-start">Description</td>
-                        <td t-if="has_option_discount" name="th_option_discount" groups="product.group_discount_per_so_line" class="text-start">Disc.%</td>
+                        <td t-if="has_option_discount" name="th_option_discount" class="text-start">Disc.%</td>
                         <td name="th_option_price_unit" class="text-end">Unit Price</td>
                     </tr>
                 </thead>
@@ -21,7 +21,7 @@
                         <td name="td_option_name">
                             <span t-field="option.name">Extra Chips</span>
                         </td>
-                        <td t-if="has_option_discount" name="td_option_discount" groups="product.group_discount_per_so_line">
+                        <td t-if="has_option_discount" name="td_option_discount">
                             <strong t-if="option.discount != 0.0" class="text-info">
                                 <t t-out="((option.discount % 1) and '%s' or '%d') % option.discount">-</t>%
                             </strong>
@@ -41,7 +41,7 @@
                 </tbody>
             </table>
             <div class="oe_structure">&#8203;</div>
-        </div>           
+        </div>
     </xpath>
 </template>
 </odoo>
